requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
django-debug-toolbar==3.2.4
djangorestframework==3.15.2
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pep8-naming==0.13.2
platformdirs==4.3.6
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
